Grid from string and Field from Grid
